pyproject.toml
[project]
name = "soc-sage"
version = "0.1.0"
description = ""
authors = [
    {name = "Itsuki Urakawa",email = "uraitsu2000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = true
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-adk (>=1.3.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "soc_sage", from = "soc_sage"}]


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
pyink = "^24.10.1"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
